create Teachable module, refactor teacher classes to include
