+++ senior_teacher.rb
@@ -1,46 +1,21 @@
 require_relative 'person.rb'
+require_relative 'teachable.rb'
 
 class SeniorTeacher < Person
+
+  include Teachable
+
   attr_reader :salary, :performance_rating, :target_raise
 
   def initialize(options={})
     super 
     @phase = 3
     @target_raise = 1000
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
+    @required_rating = 90
   end
 
   def teach_stuff
-    response = ""
-    response += "Listen, class, this is how everything works, fo SHO! "
-    response += "*drops flat-out insane knowledge bomb* "
-    response += "... You're welcome. *saunters away*"
-    response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 90
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
-    response
+    resonse = "Listen, class, this is how everything works, fo SHO! *drops flat-out insane knowledge bomb* ... You're welcome. *saunters away*"
   end
 
   def lead_training_session
