+++ teachable.rb
@@ -1,0 +1,26 @@
+module Teachable
+
+  def set_phase(num)
+    @phase = num
+    "Cool, I've always wanted to teach phase #{num}!"
+  end
+
+  def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+
+  def set_performance_rating(rating)
+    if rating > @required_rating
+        receive_raise(@target_raise)
+        response = "Yay, I'm a great employee!"
+    else
+      response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
+    end
+  end
+
+end
